fix(FeatureGroupClient): BatchGetRecord limit (#1018)
The Sagemaker Developer Guide clearly states that this limit is 100:

> BatchGetRecord API: Can contain as many as 100 records and can query up to 10 feature groups.

What had me confused is that the BatchGetRecord API documentation says that the Identifier argument can have a "Minimum number of 1 item. Maximum number of 10 items." However, an "item" in this context is a BatchGetRecordIdentifier object that can take an array of 100 record identifiers.